:package: Feat: Init eth-trpc package (#734)
## Description

EVMts runs both in the client and on the server. To make this work well
we are creating a generic TRPC interface. There will be three ways to
talk to the EVMts vm with this api

1. CreateCaller for when you are directly talking to the client
2. TRPC server if you want to run TRPC on the backend
3. Talk to TRPC in service worker (guessing this will use CreateCaller
and implement service worker from scratch but later we may be able to
generalize it)

Currently only added call because it has no generics involved. Getting
generics to work in TRPC for contractCall is the next challenge

## Testing

Explain the quality checks that have been done on the code changes

## Additional Information

- [ ] I read the [contributing docs](../docs/contributing.md) (if this
is your first contribution)

Your ENS/address:

---------

Co-authored-by: Will Cory <willcory@Wills-MacBook-Pro.local>